BSC_Tests: Avoid race condition with OML
We use the "oml-connection-status == degraded" to determine if
the OML connection has been brought up by the external helper.

However, this status is advertised from the moment the ID_ACK was
issued on the OML connection, and *not* only from the moment on
when we're expecting the RSL connection (after initializing all MO).

Let's introduce some delay to avoid running into a race condition here.

The proper solution is to implement OML inside the test suite, which
is too large of a task to be done right now, where the focus is on
RSL+BSSAP testing.